requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
Django==3.0
django-mssql-backend==2.8.1
djangorestframework==3.12.2
idna==2.10
msal==1.8.0
mysqlclient==2.0.2
pycparser==2.20
PyJWT==1.7.1
pyodbc==4.0.30
pytz==2020.4
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
